+++ minhas-financas-back-end/MinhasFinancas.Domain/Services/AnaliseFinanceira/Modelos/PainelIndicadoresFinanceiros.cs
@@ -8,6 +8,9 @@ namespace MinhasFinancas.Domain.Services.AnaliseFinanceira.Modelos
         public IndicadorFinanceiro ReservaEmergenciaIdeal { get; set; } = new();
         public IndicadorFinanceiro ComprometimentoRenda { get; set; } = new();
         public IndicadorFinanceiro ComprometimentoFinanceiroFuturo { get; set; } = new();
+        public IndicadorFinanceiro ComprometimentoFinanceiroFuturo90Dias { get; set; } = new();
+        public IndicadorFinanceiro ComprometimentoFinanceiroFuturo180Dias { get; set; } = new();
+        public IndicadorFinanceiro ComprometimentoFinanceiroFuturo365Dias { get; set; } = new();
         public IndicadorFinanceiro Endividamento { get; set; } = new();
         public IndicadorFinanceiro PatrimonioLiquidoAtual { get; set; } = new();
         public IndicadorFinanceiro PercentualPatrimonioAlvo { get; set; } = new();